+++ source/raw/bin/x64/plugins/cyber_engine_tweaks/mods/Character Preset Manager (CET)/modules/state.lua
@@ -134,6 +134,7 @@ state = {
       advancedDiagnostics = false,
     },
     helpOpen = false,
+    helpSearchText = "",
     debugOpen = false,
     settingsOpen = false,
     debugLogText = "",